Refactor de Element generation setup part
