+++ docs/manuals/gui/corelements/index.md_template
@@ -1,0 +1,11 @@
+# Core elements
+
+Taipy provides a set of visual elements that are meant to simplify the use of Core
+objects... TODO
+
+## Elements list
+
+Here is the list of all the available Core elements in Taipy:
+
+[TOC]
+
