+++ docs/manuals/gui/corelements/scenario_selector.md_template
@@ -1,0 +1,6 @@
+Shows all the scenario entities handled by Taipy Core, and lets the user
+select them.
+
+## Usage
+
+TODO
