+++ tools/_setup_generation/setup.py
@@ -104,6 +104,7 @@ class SetupStep(ABC):
 
 
 from .step_viselements import VisElementsStep
+from .step_corelements import CoreElementsStep
 from .step_refman import RefManStep
 from .step_rest_refman import RestRefManStep
 from .step_gui_ext_refman import GuiExtRefManStep
@@ -115,6 +116,7 @@ def run_setup(root_dir: str, steps: List[SetupStep] = None):
     if steps is None:
         steps = [
             VisElementsStep(),
+            CoreElementsStep(),
             RefManStep(),
             RestRefManStep(),
             GuiExtRefManStep(),
